got userName lookup on every query working; next, add restrictions based on permissions/whether or not user actually exists

+++ server/datasources/models.js
@@ -50,9 +50,32 @@ class ModelAPI extends DataSource {
     this.context = config.context;
   }
 
+  async getUser() {
+    const { user } = this.context;
+    console.log('user is');
+    console.log(user);
+    if (user !== null) {
+      const exist = await this.store.users.findAll({ where: { userName: user.userName } })
+        .then((val) => {
+          if (val && val[0]) {
+            console.log('value = ');
+            console.log(val[0].dataValues);
+            return val[0].dataValues;
+          }
+          return null;
+        });
+      return exist;
+    }
+    return null; // user === null means someone tried to access page without logging in
+  }
+
   async countAllModels() {
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return 0;
+    }
     let total = await this.store.models.findAndCountAll();
     total = total.count;
     return total;
@@ -62,6 +85,10 @@ class ModelAPI extends DataSource {
     const response = { message: '', success: false };
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return JSON.stringify(response);
+    }
     const model = await this.getModel({ modelNumber, vendor });
     const modelReference = model.dataValues.id;
     await this.store.instruments.findAll({ where: { modelReference } }).then(async (data) => {
@@ -87,9 +114,13 @@ class ModelAPI extends DataSource {
     supportLoadBankCalibration,
     categories,
   }) {
+    const response = { message: '', success: false };
     const storeModel = await this.store;
     this.store = storeModel;
-    const response = { message: '', success: false };
+    const user = await this.getUser();
+    if (!user) {
+      return JSON.stringify(response);
+    }
     const validation = validateModel({
       modelNumber, vendor, description, comment,
     });
@@ -147,6 +178,10 @@ class ModelAPI extends DataSource {
   async getAllModels({ limit = null, offset = null }) {
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return [];
+    }
     const models = await this.store.models.findAll({
       limit,
       offset,
@@ -162,11 +197,13 @@ class ModelAPI extends DataSource {
   async getModelsWithFilter({
     vendor, modelNumber, description, categories, limit = null, offset = null,
   }) {
-    console.log('******************* MODEL CONTEXT *******************');
-    console.log(this.context);
+    const response = { models: [], total: 0 };
     const storeModel = await this.store;
     this.store = storeModel;
-    const response = { models: [], total: 0 };
+    const user = await this.getUser();
+    if (!user) {
+      return response;
+    }
     if (categories) {
       let includeData;
       if (categories) {
@@ -317,6 +354,10 @@ class ModelAPI extends DataSource {
   async getAllModelsWithModelNum({ modelNumber }) {
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return [];
+    }
     const models = await this.store.models.findAll({ where: { modelNumber } });
     return models;
   }
@@ -324,6 +365,10 @@ class ModelAPI extends DataSource {
   async getAllModelsWithVendor({ vendor }) {
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return [];
+    }
     const models = await this.store.models.findAll({ where: { vendor } });
     return models;
   }
@@ -331,6 +376,10 @@ class ModelAPI extends DataSource {
   async getUniqueVendors() {
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return [];
+    }
     const models = await this.store.models.findAll({ attributes: [[SQL.fn('DISTINCT', SQL.col('vendor')), 'vendor']] });
     return models;
   }
@@ -338,6 +387,10 @@ class ModelAPI extends DataSource {
   async getModel({ modelNumber, vendor }) {
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return null;
+    }
     const model = await this.store.models.findAll({
       where: { modelNumber, vendor },
       include: {
@@ -364,6 +417,10 @@ class ModelAPI extends DataSource {
     const response = { message: '', success: false };
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return JSON.stringify(response);
+    }
     const validation = validateModel({
       modelNumber, vendor, description, comment,
     });
@@ -402,6 +459,10 @@ class ModelAPI extends DataSource {
     }
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return JSON.stringify(response);
+    }
     await this.getModelCategory({ name }).then((value) => {
       if (value) {
         response.message = `ERROR: cannot add model category ${name}, it already exists!`;
@@ -420,6 +481,10 @@ class ModelAPI extends DataSource {
     const response = { message: '', success: false };
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return JSON.stringify(response);
+    }
     await this.getModelCategory({ name }).then((value) => {
       if (value) {
         this.store.modelCategories.destroy({
@@ -450,6 +515,10 @@ class ModelAPI extends DataSource {
     }
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return JSON.stringify(response);
+    }
     let name = currentName;
     await this.getModelCategory({ name }).then(async (value) => {
       if (value) {
@@ -481,6 +550,10 @@ class ModelAPI extends DataSource {
     const response = { message: '', success: false };
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return JSON.stringify(response);
+    }
     await this.getModel({ modelNumber, vendor }).then(async (value) => {
       if (value) {
         const name = category;
@@ -509,6 +582,10 @@ class ModelAPI extends DataSource {
     const response = { message: '', success: false };
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return JSON.stringify(response);
+    }
     await this.getModel({ modelNumber, vendor }).then(async (value) => {
       if (value) {
         const name = category;
@@ -548,6 +625,10 @@ class ModelAPI extends DataSource {
   async getModelCategory({ name }) {
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return null;
+    }
     const category = await this.store.modelCategories.findAll({
       where: { name },
     });
@@ -560,18 +641,30 @@ class ModelAPI extends DataSource {
   async getAllModelCategories({ limit = null, offset = null }) {
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return [];
+    }
     return await this.store.modelCategories.findAll({ limit, offset });
   }
 
   async countModelCategories() {
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return 0;
+    }
     return await this.store.modelCategories.count();
   }
 
   async countModelsAttachedToCategory({ name }) {
     const storeModel = await this.store;
     this.store = storeModel;
+    const user = await this.getUser();
+    if (!user) {
+      return 0;
+    }
     let response = 0;
     await this.getModelCategory({ name }).then(async (result) => {
       if (result) {
